+++ python_modules/dagster/dagster/_core/definitions/partition.py
@@ -38,10 +38,10 @@ from dagster._core.instance import DagsterInstance, DynamicPartitionsStore
 from dagster._core.storage.tags import PARTITION_NAME_TAG, PARTITION_SET_TAG
 from dagster._serdes import whitelist_for_serdes
 from dagster._utils import xor
-from dagster._utils.backcompat import (
+from dagster._utils.cached_method import cached_method
+from dagster._utils.warnings import (
     normalize_renamed_param,
 )
-from dagster._utils.cached_method import cached_method
 
 from ..errors import (
     DagsterInvalidDefinitionError,
