+++ python_modules/dagster/dagster/_core/definitions/resolved_asset_deps.py
@@ -3,7 +3,7 @@ from typing import AbstractSet, Dict, Iterable, List, Mapping, Tuple, cast
 
 from dagster._core.definitions.events import AssetKey
 from dagster._core.errors import DagsterInvalidDefinitionError
-from dagster._utils.backcompat import experimental_warning
+from dagster._utils.warnings import experimental_warning
 
 from .assets import AssetsDefinition
 from .source_asset import SourceAsset
