+++ python_modules/dagster/dagster/_core/instance/__init__.py
@@ -66,12 +66,12 @@ from dagster._core.storage.tags import (
 from dagster._serdes import ConfigurableClass
 from dagster._seven import get_current_datetime_in_utc
 from dagster._utils import PrintFn, traced
-from dagster._utils.backcompat import (
+from dagster._utils.error import serializable_error_info_from_exc_info
+from dagster._utils.merger import merge_dicts
+from dagster._utils.warnings import (
     deprecation_warning,
     experimental_warning,
 )
-from dagster._utils.error import serializable_error_info_from_exc_info
-from dagster._utils.merger import merge_dicts
 
 from .config import (
     DAGSTER_CONFIG_YAML_FILENAME,
