+++ python_modules/libraries/dagster-managed-elements/dagster_managed_elements/cli.py
@@ -7,7 +7,7 @@ from typing import Optional, Sequence
 
 import click
 import click_spinner
-from dagster._utils.backcompat import disable_dagster_warnings
+from dagster._utils.warnings import disable_dagster_warnings
 
 from dagster_managed_elements.types import ManagedElementDiff, ManagedElementReconciler
