+++ python_modules/libraries/dagster-msteams/dagster_msteams/hooks.py
@@ -3,7 +3,7 @@ from typing import Callable, Optional
 from dagster._annotations import deprecated_param
 from dagster._core.definitions import failure_hook, success_hook
 from dagster._core.execution.context.hook import HookContext
-from dagster._utils.backcompat import normalize_renamed_param
+from dagster._utils.warnings import normalize_renamed_param
 
 from dagster_msteams.card import Card
