+++ python_modules/libraries/dagster-msteams/dagster_msteams/sensors.py
@@ -8,7 +8,7 @@ from dagster._core.definitions.run_status_sensor_definition import (
     run_failure_sensor,
 )
 from dagster._core.definitions.unresolved_asset_job_definition import UnresolvedAssetJobDefinition
-from dagster._utils.backcompat import normalize_renamed_param
+from dagster._utils.warnings import normalize_renamed_param
 
 from dagster_msteams.card import Card
 from dagster_msteams.client import TeamsClient
