+++ python_modules/libraries/dagster-pagerduty/dagster_pagerduty/hooks.py
@@ -4,7 +4,7 @@ from dagster._annotations import deprecated_param
 from dagster._core.definitions import failure_hook
 from dagster._core.definitions.hook_definition import HookDefinition
 from dagster._core.execution.context.hook import HookContext
-from dagster._utils.backcompat import normalize_renamed_param
+from dagster._utils.warnings import normalize_renamed_param
 
 
 def _default_summary_fn(context: HookContext) -> str:
