+++ python_modules/libraries/dagster-pagerduty/dagster_pagerduty/resources.py
@@ -4,7 +4,7 @@ import pypd
 from dagster import ConfigurableResource, resource
 from dagster._config.pythonic_config import infer_schema_from_config_class
 from dagster._core.definitions.resource_definition import dagster_maintained_resource
-from dagster._utils.backcompat import suppress_dagster_warnings
+from dagster._utils.warnings import suppress_dagster_warnings
 from pydantic import Field as PyField
